requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
greenlet==2.0.2
google-generativeai==0.4.1
requests==2.31.0
beautifulsoup4==4.12.3
cloudscraper==1.2.71
feedparser==6.0.11
PyGithub==2.1.1
python-dotenv==1.0.0
aiohttp==3.9.1
lxml==5.1.0
